data-pipeline/requirements.txt
pinecone
sentence-transformers
datasets
pandas
tqdm
python-dotenv
